Add tests Entry002 and Entry002-Bis
Entry002:
An entry point without outgoing edges that is placed on composite state
with orthogonal regions act as a fork.

Entry002-Bis:
An entry point may have outgoing transitions. If used in the context of
a composite state that has orthogonal regions, then it can have as many
outgoing transitions than the number of regions. Regions that have an
internal state targeted are entered explicitly. 

Change-Id: Ib0a1dd022501fc8ba60e8f3baef24c94d2d9ad5e
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>