Bug 469188: [Editor] Page management improvements in the Papyrus multi-editor https://bugs.eclipse.org/bugs/show_bug.cgi?id=469188
This is a squashed commit comprising all of the changes on the
committers/cdamus/469188-editor-welcome branch for the
introduction of a Welcome Page in the Papyrus Editor.

commit 51d64a198c2069b208b5231212a5514441514c54

Implement drag-and-drop rearrangement of the sections in the simple
grid-wise layout of the Welcome Page.  Persist the user's customization
in the sash/di resource with the sash model, using new classes in the
Welcome package.

commit b59f057cf1e08b51942aac1f45226a304c48039e

Fix enablement condition for Show Welcome Page action which has been
broken since the page was changed from an editor to a plain composite.

commit 69cee20b44c699c6a078e66bf9b16a83c1de7315

Remove the databinding properties APIs for Forms widgets
that are no longer needed since the NatTable refactoring
of the Diagrams table in the welcome page.

commit 00cfdd106421c0b1366c2a7deade819b4a3041ef

Add table of languages to the General section in the welcome page.

Add section of hyperlinks to related views to the welcome page.

Implement ordering of the contents of the welcome page according to
tab 'afterTab' relationships and priorities.  Overall the layout is still two rows with a variable number of columns.

commit a7d739b8e03bea5d4cf5e4d0d3663adbc62f5018

Let the SashModel report to listeners when its storage mode changes so
that the two different UI presentations of that mode can remain up-to-
date at all times.

commit 7fc4f8bf1f0fcee31f84b976e29cd7b3848ec47d

Ensure that new models initially track the active page of each tab
folder because they store their layout privately by default.

commit 6d5dbf27a4428645655bcf6ff9a99d4a3c77dcb8

Implement optional tracking of the active page in the private sash
model and switching between private and shared storage of the layout.
Includes:

* enhancement of the Sash DI model to include a currently selected
  page for each tab folder and a boolean flag whether to restore
  the last selection in each tab folder

* a new editor lifecycle event for "pre display" to provide a hook
  to set up the page activation before the SWT controls are rendered,
  so that the editor doesn't open up and then change page selections
  while the user watches

* expose the SaveLayoutBeforeClose behaviour for re-use when changing
  layout options in the welcome page

* update enablement from editable-ness for Properties widgets that
  force refresh

commit 97897b3d0d295a71a7173a584ceda3ac24c5c6ea

Factor out some more common bits into core plug-ins (e.g., the
IsWelcomeElement constraint in the properties environment).

Create the list of tables in the welcome page.  This requires some
renaming of properties model content in the diagrams list for
disambiguation.

commit f73fe34e87c915cc603ec12c608c020406ea7d73

Refactor the Diagrams table as a reusable NatTable of hyperlinks.

commit a6308f25a8c6e6c129174e9c418cb90e6f025e16

Move the column headers of the diagrams table out into a
separate composite that does not scroll vertically with
the table of hyperlinks (though it does track the table's
horizontal scroll position).

commit 0425bf34630e4420db3674fa8d4641820a305478

Add a separate preference page for Welcome Page Content.  This depends
on a new extension in the context registry for association of contexts
with preferences pages (by default, associating legacy contexts with
the Properties View preference page).

commit 06719c107a3f323be53ca6ada4f820f2d4ef10cb

Rework the contribution of welcome-page content to:

* eliminate the custom extension point for section providers
* use the Properties View framework to obtain welcome-page sections
  for the Welcome model element

Update the Diagrams table provider as a property section for the
Welcome model.  This involves implementation of a custom table-of-
hyperlinks widget.  Several reusable data-bindings classes are
added to the appropriate Papyrus infra-layer bundles.

commit 4069f62cc89413b8739cc6c33211ac5a11942bc4

Welcome page content for diagrams.

commit 8599ee6bdf7b2c7423811fee3e2441a618621ce4

The Welcome Page doesn't need to be an editor, just a composite.
So, let any page part (including simple composites) be optionally
closeable and change the WelcomePage into a simple composite part
instead of a nested workbench editor-part.

commit 83b581ea9aadb5b852b36b2fa0524ca7df60b255

Initial definition of an extension point for welcome section providers.

commit f949eaf6cf131449f4584b71072033b22cd1cea2

Initial implementation of a (so far, empty) Welcome page that is always
open when no other Papyrus editor pages are open.

Change-Id: I351118dc0b6a0dea505719a8c62ef430c273c7cf