Add test for requirement Behavior004
The execution of a doActivity Behavior of a State is not affected by the
firing of an internal Transition of that State.

Add test for requirement Encaps001

Entry points represent termination points (sources) for incoming
Transitions and origination points (targets) for Transitions that
terminate on some internal Vertex of the composite State. In effect, the
latter is a continuation of the external incoming Transition, with the
proviso that the execution of the entry Behavior of the composite State
(if defined) occurs between the effect Behavior of the incoming
Transition and the effect Behavior of the outgoing Transition. (p.325)

Add test for requirement Encaps002

If there is no outgoing Transition inside the composite State, then the
incoming Transition simply performs a default State entry. (p.325)

Change-Id: I91c087c3a4f7a866eac14a1cdb6797f66e92ac00
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>