Add test for requirement Choice001
the guard Constraints on all outgoing Transitions are evaluated
dynamically, when the compound transition traversal reaches this
Pseudostate (p.327)

Add test for requirement Choice002

If more than one guard evaluates to true, one of the corresponding
Transitions is selected. The algorithm for making this selection is not
defined. (p.327)

Add implementation for the choice pseudostate semantics

Change-Id: I1a41027c5a47722c768e7e1c410f6d1261751be6
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>