Bug 481020: [Table][requirement table] NPE during the synchronization processus https://bugs.eclipse.org/bugs/show_bug.cgi?id=481020
The problem was available when a requirement contains requirement in
tree table (the configuration must have 'packagedElement' as first depth
and 'nestedClassifier' as second depth) and the requirements was
deleted.

This problem is corrected by a test of the deletion of non-existing tree
item axis.


By this correction, an other linked bug appeared :
When the requirement was deleted and an undo was done, the requirement
was displayed twice in the table.

This was corrected by a test in the 'manageAddNotification' method by
checking that the element was not already available in the managed
elements (in the case of stereotyped element, the 'SET' event is caught
before the 'ADD' and was already managed)

+ Add the JUnit test

Change-Id: Ib9799572e43f0f7f35c54ef264d00e80f4da5bc2
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@all4tec.net>