Bug 484580 - doActivity is now started asynchronously.
Add test Behavior003 to illustrate how a doActivity is aborted when the
state is exited meanwhile the behavior is still running.

Add test Behavior003-bis to illustrate that while the doActivity is not
finished then we remain on the same state.

Add test Behavior004 to illustrate that an executed doActivity is not
affected by firing of an internal transition.

Change-Id: I649d6420ff4f393fda0c3d1da3094c7a0aad7b54
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>