Bug 485214: [Tests] Memory leak in model validation https://bugs.eclipse.org/bugs/show_bug.cgi?id=485214
Ensure that the Properties view doesn't leak the model by activating
it again before closing the editor so that the Model Validation view
isn't still active when the editor is closed.

Also fix a bug in the Duck utility that caused a ClassCastException
bombing a table-editor test in the dev environment (but not in the
build, oddly).