Bug 485220: [Architecture] Provide a more modular architecture https://bugs.eclipse.org/bugs/show_bug.cgi?id=485220
Fix regression caused by removal of the EMF Facet content-provider
handling in the EMFHelper::getEObject API.  The facet tree nodes
don't provide the object that we need as their EObject adapter,
so we introduce a new generalized service for resolving EObjects.

Also fix the failure to run the Profile Drafter tests in the build,
which aren't in a fragment bundle but actually a plug-in bundle.