Bug 486834: Language service doesn't understand new resources https://bugs.eclipse.org/bugs/show_bug.cgi?id=486834
A resource that is new (such as in the New Model Wizard) cannot provide
a content-type via its persisted state because that doesn't yet exist,
so it is necessary to infer (as much as possible) a content-type from
the in-memory state.  This also introduces a cache by schema that can
help to avoid the cost of accessing persisted content redundantly for
the purpose of descibing the content.