Bug 485220: [Architecture] Provide a more modular architecture https://bugs.eclipse.org/bugs/show_bug.cgi?id=485220
Factor UI dependencies out of the UML Element Types bundle.  This includes
moving some advices that interact with the user into a new
org.eclipse.papyrus.uml.service.types.ui bundle.

Pull up the PasteCommandService and IPasteCommandProvider API into the
Infra Diagram layer where the extension point is defined.  Deprecate the
old API in the UML layer.

Introduce a service for participation of languages in CSS styling:

* styling reset actions in the Reset Style command
* access to semantic model classes and properties to make available to CSS

Factor PapyrusObservableValue and cohorts out of the UML Tools bundle
into the Infra Layer for more general reuse and to relieve the
Diagram Infrastructure layer of UML dependencies.  The old API remains as
deprecated.

Remove the Infra Diagram Layer dependency on UML Layer for property
testers governing deletion in the diagram.  Includes introduction of
a new IGraphicalDeletionHelper OSGi service for delegation of the
determination of whether an element can be deleted from the diagram and
replacement of the XML expression properties

* org.eclipse.papyrus.uml.diagram.common.isSemanticDeletion
* org.eclipse.papyrus.uml.diagram.common.isReadOnly

by

* org.eclipse.papyrus.infra.gmfdiag.common.isSemanticDeletion
* org.eclipse.papyrus.infra.gmfdiag.common.canDelete

(where the latter is the negation of the property that it supersedes)

Extract UML dependencies from the Diagram Outline and CSS Editor bundles.

Remove unused MDTUtil APIs that referenced a UML-specific annotation.

Move the Diagram Infrastructure CSS Palette bundle into the UML layer because it
serves to provide extensions on the Palette Service, which is an overtly
UML-specific capability.

All client APIs for the Properties View are moved from
org.eclipse.papyrus.views.properties bundle to a new
org.eclipse.papyrus.infra.properties.ui bundle.  This
includes renaming of:

* extension points
* label-provider contexts
* XWT namespaces

Add an "all UI tests" suite.

Define a componentized hierarchical build layout of the main plug-ins

Change-Id: I43f8f3644857a18b69715f5a2f1da9b1cf286d67