Bug 481835: [Table] java.lang.IndexOutOfBoundsException: Index: 0, Size: 0 with UMLReferenceTextWithCompletionCellEditor https://bugs.eclipse.org/bugs/show_bug.cgi?id=481835
If the object to edit has a same namespace and shortest qualified name,
use the simple reference selection dialog instead of text completion.

It is possible to choose class with same name by the xtext (and the
completion (which displayed all objects names)) and the element choose
is the first.

Change-Id: I35e31fc6612cef5e7827638315d33d23da5e2763
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@all4tec.net>