Bug 474467: Papyrus editors title do not update when their label change
Add support for propagation of label-provider change events for labels
that are dependent on other elements' labels.  Implement such a
dependent label for all notation views supported by the Viewpoints
mechanism that have no names of their own, including

* unnamed Diagrams
* unnamed Tables

Now that the table label provider depends on table prototypes,
the broken prototype references in editor reload tests need to
be updated (they were missed in the refactoring when all of the
prototypes were removed from the builtin Viewpoints configuration
model).

Change-Id: I8a9c361129c996188f87ac2851db39e0f66f3acd