Bug 468207: [StateMachine Diagram] cannot delete elements https://bugs.eclipse.org/bugs/show_bug.cgi?id=468207
When a state machine has only one region, that region's zone annotation
is the empty string.  When that region has a stereotype applied (as
it does in Papyrus-RT real-time state machines), then the region has
a stereotype-comment compartment.  In that case, deletion of the
region view fails with a SIOOBE on attempt to take a substring of the
empty zone annotation value.