Bug 490936: New Model Wizard always creates a model of the kind selected in previous invocation https://bugs.eclipse.org/bugs/show_bug.cgi?id=490936
Ensure that a wizard allowing only a single diagram category only
selects one, so that the wrong kind of model isn't created because
the user's last selection happened not to be the first in the list
of diagram categories.

Also only save the selected category in the dialog settings when
finishing the wizard; cancel should not remember anything from the
dialog session.

Change-Id: Iae567c4a0b3bea579f9256eef3047bd7ac16bd09