Bug 491542: [All Diagrams] Undo doesn't work with RuntimeValuesAdvice dialog for creation https://bugs.eclipse.org/bugs/show_bug.cgi?id=491542
Support re-entrant command execution in the CheckedOperationHistory,
which is used by both the CheckedDiagramCommandStack (used by the
diagrams) and the NestingNotifyingWorkspaceCommandStack (used by
everything else), in much the same way as the latter supports
nested execution.  This ensures that, in the case where the execution
of a top-level operation in the diagram triggers nested execution of
operations in the EMF editing domain, which both delegate to the same
CheckedOperationHistory, the nested operations are executed strictly
within the context of the root operation, not separately as a sequence
of operations in the history.

Change-Id: I0967a62940d5c1030edf14d39994eeda2d3bdf9b