Bug 493853: [Properties] NattablePropertyEditor illegally modifies its configuration model https://bugs.eclipse.org/bugs/show_bug.cgi?id=493853
Simply don't attempt to change the axis configuration in the table
configuration.  There is no point anyways, because once the table
instance has been created and saved, we will never create it again
and so in subsequent sessions we wouldn't be poking the table
configuration again.  Besides that the table configuration is a
read-only model deployed in a plug-in bundle.

Change-Id: If06162c928248ec4b9fabf30144c0449c6885bae