Bug 494478: StereotypeElementListener dispatches events too early during undo/redo https://bugs.eclipse.org/bugs/show_bug.cgi?id=494478
Restore the post-commit timing of the dispatch of custom notifications for stereotype applications generated by undo and redo of commands.
This ensures that, for example, Adapters attached directly to objects
are not notified of unapplication of a stereotype during undo before
that stereotype is actually unapplied.

The pre-commit timing of the original notifications from command
execution is preserved because:

* they do not have the problem of preceding the actual model changes
* pre-commit listeners such as applied-stereotype display edit-policies
  still need to receive and react to these notifications before commit
* post-commit listeners will still receive these also, as before

Delaying the notifications to post-commit phase for undo/redo should not break any listeners' timing assumptions because only post-commit listeners could have received them, anyways:  undo/redo transactions do not have the pre-commit (triggers) phase.

Change-Id: Iaeda2abfbb86d98fbdf0a93787bef25901a7e312