Bug 495312: [Table 2] Reordering model-elements duplicates them in hierarchical synced table https://bugs.eclipse.org/bugs/show_bug.cgi?id=495312
The rows duplicated was managed by an add object into the table even if
the row of element was already in the table.

For the initial order, this is managed as specified in the 'node'
attribute of the 'activity' node (and not like sub elements of it (as
model explorer did)).

Care to your used feature: For activity nodes, this are displayes as
ownedElements and as ownedNodes (2 differents attributes).

Change-Id: I51c7b593d4cb91b97f71a0110325a731e376a596
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@all4tec.net>