Bug 495312: [Table 2] Reordering model-elements duplicates them in hierarchical synced table https://bugs.eclipse.org/bugs/show_bug.cgi?id=495312
The rows duplicated was managed by an add object into the table even if
the row of element was already in the table. 

For the initial order, this is managed as specified in the 'node'
attribute of the 'activity' node (and not like sub elements of it (as
model explorer did)).

Care to your used feature: For activity nodes, this are displayes as
ownedElements and as ownedNodes (2 differents attributes).

Change-Id: Ib0a8d69565691a385c735a44e26c5d04f799d05e
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@all4tec.net>