Bug 496439: [RSA Import] Give mapping traces to extension transformations https://bugs.eclipse.org/bugs/show_bug.cgi?id=496439
Carry mapping traces forward from each transformation in the chain
to the next.  Provide a black-box library with custom operations that
query the mapping traces to find sources for generated outputs because
the standard invresolve and invresolveone operations only inspect the
traces of the current transformation, not any previous.

Provide also a new API for execution of a transformation that
automatically includes the traces, so that extensions can make use
of them, too (otherwise, as currently in Papyrus-RT, they will always
have to reimplement access to the executor pool, which is a detail
that should be abstracted away from them).

Change-Id: Ibeece23d01bf8fb86bf714995141cfc01685a509