Bug 496307: [Profile Migration] Restrictive package pattern matching
- Add ProfileNamespaceURIPatterns to split namespace URIs into a
versionless, identifying namespace URI and the version information. 
A namespace URI is matched by a given Regex pattern. The versionless
namespace URI is a concatenation of all parts of the matches that are
not in groups. The version information is a formatted string consisting
of the parts matched through the groups in the pattern and therefore may
be spread across the URI. The format of the  resulting version string
can be customized using a dedicated MessageFormat.
- Add extension point for profile namespace URI patterns.
- Add pattern registry that is initialized with extension points.
- Add usage of profile namespace URI patterns when comparing URIs in the
zombie stereotypes descriptor, if there is no perfect equality. If no
registered pattern handles the URIs, the previous behavior is used as
default.
- Bump plugin version to 1.2.1

Includes tests for the pattern matching and splitting.
Includes tests to test reported behavior and the improved behavior in
the zombie stereotypes descriptor.

Change-Id: I5f7744c2a19bb20ea2572247f6d47f948504f7b4
Signed-off-by: Martin Fleck <mfleck@eclipsesource.com>