Bug 495394: Reveal first match according to text in filter
The EMFGraphicalContentProvider will now find and reveal the first
element that matches the given filter. To avoid bad performance when
searching for the first matching element, we "just" unfold the first
matching item in the already filtered tree viewer. Even for large models
this should be quick, as it will be only take the cost of getting the
matching children (by expanding the parent) for n times, whereas n is
the depth of the tree to the first leaf element. To avoid stackoverflow
errors in infinite tree, we use a maximum depth cap of 15.

Change-Id: Ie15b67366407423049c9735032b94cb2131b657e
Signed-off-by: Philip Langer <planger@eclipsesource.com>