Bug 495394: Hide description in TreeSelectorDialog if there is none
The description label produces empty space even if there is no
description, which happens in many cases, e.g., when we click the "..."
button for selecting the type in the property view of an attribute.

To avoid this empty space before the filter text box, we hide the label
(and prevent it from using space) if the description is null and only
show it when it is not null. If the description is set at a later point,
it will be still shown, as we "unhide" the label and call pack() on the
parent composite.

Change-Id: Idcab2f73db34544432e3dfb99c1ebf8e5f981ddb
Signed-off-by: Philip Langer <planger@eclipsesource.com>