Bug 497841: [Model Import] Size of imported capsule part is smaller than in legacy model causing additional layout issues https://bugs.eclipse.org/bugs/show_bug.cgi?id=497841
Don't attempt to create port views as children of the structure diagram
itself (not the frame) because they wouldn't be provided by the view
service, anyways.

Ensure proper himetric conversion to pixels according to the host
system's display resolution and the 2540 units per inch scaling to
avoid considerable drift in the layout.

(cherry-picked from streams/2.0-maintenance)
Change-Id: I4bc66c014cca85ca35c5c34e3e4ef856da54123b