Bug 498282: [Model Import] Visualization of port conjugation does not work on imported structure diagrams https://bugs.eclipse.org/bugs/show_bug.cgi?id=498282
Ensure that the default styling of conjugated ports is not overridden
by an explicit fill colour from the imported RSA diagram.

Also, add an advice on the Port::isConjugated property that, whenever
it is set true from false, any explicit fill colour in existing views
of that port is unset.

(cherry-picked from streams/2.0-maintenance)
Change-Id: I0af883058881da955c3d15ca28819f0ad9eb46e0