Bug 498359: [Index] UI deadlock in model save scenario https://bugs.eclipse.org/bugs/show_bug.cgi?id=498359
Provide the current state of the index immediately when requested while
the current thread is processing resource changes.  None of the re-index
jobs can proceed to modify the index, so it is as good and consistent a
picture of the index as can be provided and otherwise we would deadlock.

Change-Id: Ia0445d4712e6d8ef74c866e5354f308544a0670b