Bug 498282: [Model Import] Visualization of port conjugation does not work on imported structure diagrams https://bugs.eclipse.org/bugs/show_bug.cgi?id=498282
Make the import tests independent of Papyrus-RT's transformation
extension, which populates the UML 2.5 Port::isConjugated attribute
from the RSA UML-RT profile's RTPort stereotype.  Add a transformation
extension that is active only during the tests, which looks for the
AnyTypes of missing-schema RTPorts to transfer the 'isConjugate'
property to the base port.  When Papyrus-RT is installed, it will do this for us and the RTPort stereotype instances will not be AnyTypes,
so our transformation extension will have no effect.

Change-Id: I34c9a9740cad9beb890564b30b20eab16594a9cd