Bug 499661: Exception about missing write transaction when iterating over a potential unloaded model element https://bugs.eclipse.org/bugs/show_bug.cgi?id=499661
The listener for loading of resources was still causing problems,
triggering re-loading of editor pages in situations where it doesn't
need to.  Moreover, this listener was implemented in the wrong
component:  its purpose is solely to re-load an "unloaded resource"
editor tab as created by the UI for the selective resource loading
service.  Therefore, it is this UI bundle that now provides the
listener and this listener now further restricts its reload behaviour
to the case only when the editor page being reloaded is one of the
"unloaded resource" pages that it had created, itself.

To maintain API compatibility, the LoadDiagramCommand, with its
previous fixes, is retained in the control-mode bundle but is now
deprecated.  The listener snippet that used it is moved to the
resource-loading service bundle (it was not API) along with a new
implementation of the LoadDiagramCommand that is also not API.

Change-Id: I06ebcd232fd07cdc64db5a5b371940020716ac36