Bug 500046: [Control mode] Deadlock possibility at startup https://bugs.eclipse.org/bugs/show_bug.cgi?id=500046
Ensure that the Papyrus editor matching strategy can never cause
deadlock by hitting the full cross-reference index only when it is
already available.  In the case that it would be necessary to wait
for the index to be calculated, fall back to the on-demand index.

This uses a new general-purpose "ifAvailable" operation in the index
internal API, which can be used to provide similar capabilities for
the other queries (currently it is used only for the get-roots
operation).

Change-Id: I2dcd691de49d33f485dce2a4f330dbbf4c2b96b1