Bug 497865: [Control Mode] UI for dependent controlled units https://bugs.eclipse.org/bugs/show_bug.cgi?id=497865
Add an option to the standard control (create fragment) dialog for
option to create sub-model unit (independently openable) or not (in
which case it's a 'shard').

Add an approver interface to the participant protocols for validation of
control/uncontrol requests and also to determine whether an object
supports the sub-model unit form of controlled resource.  Use this to
let the UML participant disable the sub-model unit option for
non-packages.

Add a label decorator for controlled unit resources in the Project
Explorer, with different presentation for 'shards' as for sub-model
units that are independently openable.

Add a context menu on model elements to toggle their independent
sub-model status when they are currently controlled units.  This is
supported by another optional participant protocol for changing
sub-unit mode so that for UML we can add/remove redundant profile
applications on packages as necessary.

(cherry-picked from streams/2.0-maintenance)
Change-Id: I08ac9cc64d70432851f8e7e815f352f4b00d50f2