Bug 501946: Undo creation of capsule with structure diagram open causes exception to be thrown https://bugs.eclipse.org/bugs/show_bug.cgi?id=501946
Ensure that, when a notation view is removed from the notation resource,
then if any editor pages open on it are not closed deliberately, they
are closed implicitly.  Also do not attempt to refresh a diagram edit
part that has lots its semantic element, on the assumption that it is
being deleted.

Change-Id: I3840df7890c41785fbf079edd890ae29133ac2e4