Bug 501946: Undo creation of capsule with structure diagram open causes exception to be thrown https://bugs.eclipse.org/bugs/show_bug.cgi?id=501946
Ensure that, when a notation view is removed from the notation resource,
then if any editor pages open on it are not closed deliberately, they
are closed implicitly.  Also do not attempt to refresh a diagram edit
part that has lots its semantic element, on the assumption that it is
being deleted.

In order to reuse the asynchronous executor for an editing domain,
factor it out of the DiagramHelper into a more suitable common API.

(cherry-picked from streams/2.0-maintenance)

Change-Id: Ic0b59ff9b3475f8ab4e7d852b0719c9ba479758b