Bug 505330: Unable to post-process imported/migrated RSA-RTE model with proprietary profiles https://bugs.eclipse.org/bugs/show_bug.cgi?id=505330
Ensure that, as much as possible, the schema locations for profiles
are resolved when loading the model resources to be migrated, using
the same dependency helper as is used in fixing references later in
the migration process.  But only when the profile URI doesn't actually
seem to exist as is.

Includes JUnit tests with models provided by contributors for
reproduction of the original problem.  Also a bit of refactoring of
existing test infrastructure now that we have another regression
test suite.

Change-Id: Ib5fa141704566f61c3696f311ac942da0ad458d1