Bug 508629: [Label Provider Service] Label providers do not notify
Fix gaps in the propagation of label updates from EMF item provider
adapters through label-providers that delegate to them and the
label-provider service that delegates to those, in turn.

Add JUnit tests covering various affected label providers, including
a new tests plug-in for the Properties UI in the Infra layer.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=508629

Change-Id: I2e0301db72064fe6899c4d92252d93d9051943f0