Bug 508184:  [Sequence Diagram] Start and finish of execution specification must be in same interaction operand
- When Move or resize an ExecutionSpecification, testing that the future
bounds respect the rule
- If not it cannot be moved or resized
- In a semantic point of view, ExecutionSpecification , starting and
ending Occurence should now be always into the same
interaction/interaction operand. 

NOT FIXED: When the user creates a new executionSpecification next to
the bottom boundary of the InteractionOperand, the system allow him to
do so. Even if the ES seems belonging to 2 different
interaction/InteractionOperand. 

But in a semantic point of view this is not the case. 
 


Change-Id: I7a1fbd9591ef003660452dc2d9356632b0591a36
Signed-off-by: Céline Janssens <Celine.Janssens@all4tec.net>