Bug 505330: Unable to post-process imported/migrated RSA-RTE model with proprietary profiles
Add an optional transformation extension interface for a new
post-processing phase invoked on the final imported model
after stereotype repair but before it is finally saved.

Also persist the basic import configuration parameters from
either the simple dialog or the wizard in the workspace metadata
area so that they will be the defaults for the next import.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=505330
(🍒-picked from streams/2.0-maintenance)

Change-Id: I364023b4529fb46197da0702b6b52bd21c223cfd