Bug 511917 - [Model import] update the model import tool to produce a Port_NameLabel view without an explicit set of element
* the toLabel() mapping that creates the DecorationNode PortNameLabel
inherits from toPapyrusNode mapping that inherits from toPapyrusNodeView
mapping that set the  
element : PortNameLabel should not in any case set the element, the
element is set in the Port_Shape view. for the inherited ports this
creates a conflict and did not display the label -->  create an abstract
mapping just like PapyrusConnectorLabel and ensure that we did not set
the element in this mapping. toLabel() inherits now the new mapping  

* add junit test to be sure that for all PortNameLabel View no element
was set (test only simple port, inherited port should be tested in
Papyrus RT not in Papyrus)

Change-Id: I68eb807294d9461e4a8dd448638030a5c0a5a472
Signed-off-by: Asma Smaoui <asma.smaoui@cea.fr>