Bug 514916: [Test] Tests failure for class diagram https://bugs.eclipse.org/bugs/show_bug.cgi?id=514916
Modify the gmfgen file to manage
papyrusgmfgenextension:VisualIDOverride.

The element types id generated for the palette for associations need to
be not generated because, actually, the type id is managed by the edit
part. To generate the correct element types id for the palette factory,
it is needed to create some edit parts, some new visual ids, and so on.

Change-Id: I76c6e94734f39a58ee7f96e857613145eb3abcf0
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@cea.fr>