Bug 513267: [State machine diagram] Superfluous scrollbars always present https://bugs.eclipse.org/bugs/show_bug.cgi?id=513267
Initialize the height Header of StateMachine to 20 because after the
creation, the CustomStatsMachineNameEditPart refreshed it and
automatically resize it to 20.

When it was initialized to 13, the height of created region was
calculated by the height of the StateMachine minus the size of the
header (so 300 - 13 = 287). And when the header was resized to 20, the
StateMachine was resized to 280. So the Region was taller than the
StateCompartment and the ScollBar was displayed (because 287 in 280
compartment).

Change-Id: I4e5d0c1d21b03fea63ee2ae0246a6f3af63b160e
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@cea.fr>