Bug 489116: [Sequence Diagram] Migrate Duration Observation for elementtypes
- Activate Custom Duration Observation EditPart

Know ISSUE : Parser passed is the composite parser instead of Sequence
parser for DurationObservation

Solution 1 : Have the same Parser for all the DurationObservation
Issue 1 : Duration Observation Parser herit from MessageFormatParser
that is generated for each diagram. And cannot be common

Solution 2 : Modify the Element Type ID in order to identify if an
element type belongs to a specific diagram to avoid to match a parser
from another diagram.
Issue2: No idea of the side effects 

Solution 3 : Create a specific Parser Provider for DurationObservation
of the sequence diagram and put it with a higher priority
Issue 3:  I'm pretty sure, then this is the composite
DurationObservation that will match the Sequence DurationObservation
Parser. 



Change-Id: I3da0abad30617c3e8243ce0c070d84b1fab073d9
Signed-off-by: Céline Janssens <Celine.Janssens@all4tec.net>